Implementado activity para avaliação, e uma para visualizar as avaliações

+++ app/src/main/java/com/example/cadastroalunos/dao/DisciplinaDAO.java
@@ -18,7 +18,7 @@ public class DisciplinaDAO {
         }
     }
 
-    public static Disciplina getDisciplina(int id){
+    public static Disciplina getDisciplina(Long id){
         try{
             return Disciplina.findById(Disciplina.class, id);
         }catch (Exception ex){
